accounts/requirements.txt
flask
requests
pytest
black
python-dotenv